requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
python-dotenv
python-docx
werkzeug
lxml
WTForms-SQLAlchemy
gunicorn
